Jit64: Split generic and x64 helpers.
