+++ Source/Core/Core/CMakeLists.txt
@@ -199,6 +199,8 @@ if(_M_X86)
 			PowerPC/Jit64/JitRegCache.cpp
 			PowerPC/Jit64/Jit_SystemRegisters.cpp
 			PowerPC/Jit64Common/Jit64AsmCommon.cpp
+			PowerPC/Jit64Common/JitBase.cpp
+			PowerPC/Jit64Common/JitCache.cpp
 			PowerPC/JitCommon/JitBackpatch.cpp
 			PowerPC/JitCommon/Jit_Util.cpp
 			PowerPC/JitCommon/TrampolineCache.cpp)
