+++ Source/Core/Core/PowerPC/JitCommon/JitBackpatch.cpp
@@ -14,7 +14,7 @@
 #include "Common/MsgHandler.h"
 #include "Common/x64Emitter.h"
 #include "Core/HW/Memmap.h"
-#include "Core/PowerPC/JitCommon/JitBase.h"
+#include "Core/PowerPC/Jit64Common/JitBase.h"
 
 using namespace Gen;
